[PORT FROM MAIN][BT] Fix sign propagation issue due to cast to (int)
BZ: 74811

Due to integer compilation difference between Arm (unsigned)
and x86 (signed), the cast to (int) could induce a sign propagation
that would create a stack corruption.
To fix this issue, we specify precisely the cast to (UINT8),
in order to be sure of the outcome.

Change-Id: I99ed6f132d5c16e2bbacc922f2d43d2ad4f3e8d6
Orig-Change-Id: I2f1e36f672e579df06b0d9c416a3d8a4cedfdc9b
Feature: BT
Category: Integration
Purpose: PV-Enable
Origin: Internal
Upstream-candidate: yes
Signed-off-by: Cedric Bondier <cedricx.bondier@intel.com>
Reviewed-on: http://android.intel.com:8080/82859
Reviewed-by: Bransiec, ChristopheX <christophex.bransiec@intel.com>
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Biselli, MirkoX <mirkox.biselli@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>