[PORT FROM MAIN][BT] Fix sign propagation issue due to cast to (int)
BZ: 74811

Due to integer compilation difference between Arm (unsigned)
and x86 (signed), the cast to (int) could induce a sign propagation
that would create a stack corruption.
To fix this issue, we specify precisely the cast to (UINT8),
in order to be sure of the outcome.

Category: device enablement
Domain: CWS.BT-Common
Origin: internal
Upstream-Candidate: yes

Change-Id: Ic09996821fb138aae47051fdb599494925a0d8c2
Orig-Change-Id: I99ed6f132d5c16e2bbacc922f2d43d2ad4f3e8d6
Signed-off-by: Cedric Bondier <cedricx.bondier@intel.com>
Reviewed-on: http://android.intel.com:8080/82859
Reviewed-by: Bransiec, ChristopheX <christophex.bransiec@intel.com>
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Biselli, MirkoX <mirkox.biselli@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>