[PORT FROM MAIN][BT] Add -testmode option to bdt to allow use of nohup
BZ: 101809

For some implementations (like with a TI chip), the Bluetooth chip is
reset when bdt closes. Thus, when bdt is launched from an adb shell,
it is not possible to unplug the USB cable (if the device needs to be
put in test mode and then placed in a shielded box for example).

A workaround for this is to call bdt with nohup. But because bdt takes
some commands from stdin, we need an option to place it directly in
test mode at startup. Hence this additional "-testmode" option.

Calling
  bdt -testmode

is equivalent to calling
  bdt
and then launch the following commands:
  enable
  <wait for enabling>
  dut_mode_configure 1

To call bdt in an adb shell and let it run even if the shell is
closed, call:
  nohup bdt -testmode &

Category: aosp improvement
Domain: CWS.BT-Common
Origin: internal
Upstream-Candidate: yes

Change-Id: Ib10e88870b79deae74edd615840ff1e0984c7a17
Orig-Change-Id: I789c3e96e1eea696dd51a23c6e3273e1f39f42ee
Signed-off-by: Xavier Boniface <xavierx.boniface@intel.com>
Reviewed-on: http://android.intel.com:8080/104120
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>