[BT] BT crash in loop after first BLE connection
Division by 0 leads to a crash.

List of BLE devices is retrieved from controller through
a vendor specific command. The size of the list is used in a division.
When no BLE devices was previously paired, the list is empty:
need to add a zero-check in bluedroid before performing the division.

Change-Id: Id7dd2faa010d1c15489a163405ace53c65ed26bb
Orig-Change-Id: I5b4b73b4160d61bb4276c63d2bfdbbc007a0c5b9
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-187
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11039
Category: device enablement
Domain: CWS.BT
Origin: internal
Upstream-Candidate: yes
Signed-off-by: jbelloua <jean-marie.bellouard@intel.com>
Reviewed-on: https://android.intel.com:443/306445