[UPSTREAM_PENDING] Adding protection from modulo by 0 operation
In vnd/ble/vendor_ble.c an unchecked modulo
operation on a variable that sometimes can be 0,
causes a SIGPFE crash in Bluetooth Adapter.

On ARM processors this operation is ignored and
the variables remain the same. This patch does
the same for x86 processors by checking if
max_irk_list_sz is 0 or not before operation.

All other uses and modulo of this variable
are already checked.

Change-Id: I0e1958518862670f8a7d02b028b875a81cc8817b
Signed-off-by: dobreant <antonel.g.dobre@intel.com>
Reviewed-on: https://android.intel.com/280721
Reviewed-by: Davidoaia, Bogdan M <bogdan.m.davidoaia@intel.com>