Bluetooth: Removing HID information of un-bonded HID devices
If HID device is paired and un-paired immediately followed by bt-off and bt-on,
few tags are not properly cleaned up from bt_config.xml. As per the implementation,
once a HID is paired, only HID can connect to the phone/tablet, but device can't
initiate connection to HID. So when user wants to pair with the same HID, BT stack
checks those tags and ignores the request.

In this patch, we are removing all those tags if the HID device is not bonded
so that BT stack process HID request.

Issue: BMDA-1505
Category: aosp improvement
Domain: AOSP bluetooth
Origin: internal
Upstream-Candidate: yes

Change-Id: I90c2d199a0a770c9816f5caaaed1781cd148191f
Signed-off-by: Sathish Narasimman <sathishx.narasimman@intel.com>
Reviewed-on: https://android.intel.com/290532
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Debnath, Paritosh <paritosh.debnath@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>