Bluetooth: Fix HID mouse was changed to Joystick[Lollipop]
GAP: Try to retrieve CoD from NVRAM if 0

This patch tries to retreive CoD from NVRAM if CoD is 0. This is
because in scenarios when airplane mode on/off, inquiry is not
performed again on devices shown in Available devices List of
UI, so CoD is 0 during SSP information as database of stack is
not properly updated since inquiry is not performed again. Without
this patch it was observed because of wrong CoD info device icon
for HID mouse was changed to Joystick and device is not shown in
Paired List. Also some extra debug logs are added in code for
better debugging.

Change-Id: Ib93586d8b4d1cd37915907f8f33928312fb59c1d
Reviewed-on: https://android.intel.com/315066
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>
Tested-by: Kandru, Suneel <suneel.kandru@intel.com>